+++ core/transfer/src/test/java/org/eclipse/dataspaceconnector/transfer/core/provision/ProvisionManagerImplTest.java
@@ -1,5 +1,6 @@
 package org.eclipse.dataspaceconnector.transfer.core.provision;
 
+import org.eclipse.dataspaceconnector.policy.model.Policy;
 import org.eclipse.dataspaceconnector.spi.EdcException;
 import org.eclipse.dataspaceconnector.spi.monitor.Monitor;
 import org.eclipse.dataspaceconnector.spi.transfer.provision.Provisioner;
@@ -32,10 +33,12 @@ class ProvisionManagerImplTest {
     private final Provisioner provisioner = mock(Provisioner.class);
     private final Monitor monitor = mock(Monitor.class);
     private final ProvisionManagerImpl provisionManager = new ProvisionManagerImpl(monitor);
+    private Policy policy;
 
     @BeforeEach
     void setUp() {
         provisionManager.register(provisioner);
+        policy = Policy.Builder.newInstance().build();
     }
 
     @Test
@@ -44,14 +47,14 @@ class ProvisionManagerImplTest {
         var provisionResponse = ProvisionResponse.Builder.newInstance()
                 .resource(new TestProvisionedDataDestinationResource("test-resource"))
                 .build();
-        when(provisioner.provision(isA(TestResourceDefinition.class))).thenReturn(completedFuture(provisionResponse));
+        when(provisioner.provision(isA(TestResourceDefinition.class), isA(Policy.class))).thenReturn(completedFuture(provisionResponse));
         TransferProcess transferProcess = TransferProcess.Builder.newInstance()
                 .id("id")
                 .state(TransferProcessStates.REQUESTED.code())
                 .resourceManifest(ResourceManifest.Builder.newInstance().definitions(List.of(new TestResourceDefinition())).build())
                 .build();
 
-        var result = provisionManager.provision(transferProcess);
+        var result = provisionManager.provision(transferProcess, policy);
 
         assertThat(result).succeedsWithin(1, SECONDS)
                 .extracting(responses -> responses.get(0))
@@ -64,14 +67,14 @@ class ProvisionManagerImplTest {
     @Test
     void provision_should_fail_when_provisioner_throws_exception() {
         when(provisioner.canProvision(isA(TestResourceDefinition.class))).thenReturn(true);
-        when(provisioner.provision(isA(TestResourceDefinition.class))).thenThrow(new EdcException("error"));
+        when(provisioner.provision(isA(TestResourceDefinition.class), isA(Policy.class))).thenThrow(new EdcException("error"));
         TransferProcess transferProcess = TransferProcess.Builder.newInstance()
                 .id("id")
                 .state(TransferProcessStates.REQUESTED.code())
                 .resourceManifest(ResourceManifest.Builder.newInstance().definitions(List.of(new TestResourceDefinition())).build())
                 .build();
 
-        var result = provisionManager.provision(transferProcess);
+        var result = provisionManager.provision(transferProcess, policy);
 
         assertThat(result).failsWithin(1, SECONDS)
                 .withThrowableOfType(ExecutionException.class)
@@ -82,14 +85,14 @@ class ProvisionManagerImplTest {
     @Test
     void provision_should_fail_when_provisioner_fails() {
         when(provisioner.canProvision(isA(TestResourceDefinition.class))).thenReturn(true);
-        when(provisioner.provision(isA(TestResourceDefinition.class))).thenReturn(failedFuture(new EdcException("error")));
+        when(provisioner.provision(isA(TestResourceDefinition.class), isA(Policy.class))).thenReturn(failedFuture(new EdcException("error")));
         TransferProcess transferProcess = TransferProcess.Builder.newInstance()
                 .id("id")
                 .state(TransferProcessStates.REQUESTED.code())
                 .resourceManifest(ResourceManifest.Builder.newInstance().definitions(List.of(new TestResourceDefinition())).build())
                 .build();
 
-        var result = provisionManager.provision(transferProcess);
+        var result = provisionManager.provision(transferProcess, policy);
 
         assertThat(result).failsWithin(1, SECONDS)
                 .withThrowableOfType(ExecutionException.class)
@@ -103,14 +106,14 @@ class ProvisionManagerImplTest {
                 .resource(new TestProvisionedDataDestinationResource("test-resource"))
                 .build();
         when(provisioner.canDeprovision(isA(ProvisionedResource.class))).thenReturn(true);
-        when(provisioner.deprovision(isA(TestProvisionedResource.class))).thenReturn(completedFuture(deprovisionResponse));
+        when(provisioner.deprovision(isA(TestProvisionedResource.class), isA(Policy.class))).thenReturn(completedFuture(deprovisionResponse));
         TransferProcess transferProcess = TransferProcess.Builder.newInstance()
                 .id("id")
                 .state(TransferProcessStates.REQUESTED.code())
                 .provisionedResourceSet(ProvisionedResourceSet.Builder.newInstance().resources(List.of(new TestProvisionedResource())).build())
                 .build();
 
-        var result = provisionManager.deprovision(transferProcess);
+        var result = provisionManager.deprovision(transferProcess, policy);
 
         assertThat(result).succeedsWithin(1, SECONDS)
                 .extracting(responses -> responses.get(0))
@@ -123,14 +126,14 @@ class ProvisionManagerImplTest {
     @Test
     void deprovision_should_fail_when_provisioner_throws_exception() {
         when(provisioner.canDeprovision(isA(ProvisionedResource.class))).thenReturn(true);
-        when(provisioner.deprovision(isA(TestProvisionedResource.class))).thenThrow(new EdcException("error"));
+        when(provisioner.deprovision(isA(TestProvisionedResource.class), isA(Policy.class))).thenThrow(new EdcException("error"));
         TransferProcess transferProcess = TransferProcess.Builder.newInstance()
                 .id("id")
                 .state(TransferProcessStates.REQUESTED.code())
                 .provisionedResourceSet(ProvisionedResourceSet.Builder.newInstance().resources(List.of(new TestProvisionedResource())).build())
                 .build();
 
-        var result = provisionManager.deprovision(transferProcess);
+        var result = provisionManager.deprovision(transferProcess, policy);
 
         assertThat(result).failsWithin(1, SECONDS)
                 .withThrowableOfType(ExecutionException.class)
@@ -141,14 +144,14 @@ class ProvisionManagerImplTest {
     @Test
     void deprovision_should_fail_when_provisioner_fails() {
         when(provisioner.canDeprovision(isA(ProvisionedResource.class))).thenReturn(true);
-        when(provisioner.deprovision(isA(TestProvisionedResource.class))).thenReturn(failedFuture(new EdcException("error")));
+        when(provisioner.deprovision(isA(TestProvisionedResource.class), isA(Policy.class))).thenReturn(failedFuture(new EdcException("error")));
         TransferProcess transferProcess = TransferProcess.Builder.newInstance()
                 .id("id")
                 .state(TransferProcessStates.REQUESTED.code())
                 .provisionedResourceSet(ProvisionedResourceSet.Builder.newInstance().resources(List.of(new TestProvisionedResource())).build())
                 .build();
 
-        var result = provisionManager.deprovision(transferProcess);
+        var result = provisionManager.deprovision(transferProcess, policy);
 
         assertThat(result).failsWithin(1, SECONDS)
                 .withThrowableOfType(ExecutionException.class)
@@ -156,6 +159,7 @@ class ProvisionManagerImplTest {
                 .withMessageContaining("error");
     }
 
-    private static class TestProvisionedResource extends ProvisionedResource {}
+    private static class TestProvisionedResource extends ProvisionedResource {
+    }
 
 }